allocate variable functions to the category-specific variable modules
