+++ models.py
@@ -576,6 +576,8 @@ def neighborhood_vars(jobs, households, buildings):
     # print pd.Series(nodes.index).describe()
     orca.add_table("nodes_drv", nodes)
 
+    post_access_variables()
+
 
 @orca.step('gq_model')  # group quarters
 def gq_model(iter_var, tazcounts2015gq, tazcounts2020gq, tazcounts2025gq, tazcounts2030gq, tazcounts2035gq,
@@ -871,3 +873,82 @@ def _print_number_unplaced(df, fieldname="building_id"):
     """
     counts = (df[fieldname] == -1).sum()
     print "Total currently unplaced: %d" % counts
+
+
+def post_access_variables():
+
+    @orca.column('nodes_walk', 'node_r1500_acre', cache=True, cache_scope='iteration')
+    def node_r1500_acre(nodes_walk):
+        return nodes_walk['node_r1500_sqft'] / 43560
+
+    @orca.column('nodes_walk', 'ln_empden', cache=True, cache_scope='iteration')
+    def ln_empden(nodes_walk):
+        return np.log1p(nodes_walk.jobs / nodes_walk.node_r1500_acre).fillna(0)
+
+    @orca.column('nodes_walk', 'ln_popden', cache=True, cache_scope='iteration')
+    def ln_popden(nodes_walk):
+        return np.log1p(nodes_walk.population / nodes_walk.node_r1500_acre).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_high_income', cache=True, cache_scope='iteration')
+    def percent_high_income(nodes_walk):
+        return np.log1p(nodes_walk.highinc_hhs / nodes_walk.households).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_mid_income', cache=True, cache_scope='iteration')
+    def percent_mid_income(nodes_walk):
+        return np.log1p(nodes_walk.midinc_hhs / nodes_walk.households).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_low_income', cache=True, cache_scope='iteration')
+    def percent_low_income(nodes_walk):
+        return np.log1p(nodes_walk.lowinc_hhs / nodes_walk.households).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_race1', cache=True, cache_scope='iteration')
+    def percent_race1(nodes_walk):
+        return np.log1p(nodes_walk.race_1_hhs / nodes_walk.households).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_race2', cache=True, cache_scope='iteration')
+    def percent_race2(nodes_walk):
+        return np.log1p(nodes_walk.race_2_hhs / nodes_walk.households).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_race3', cache=True, cache_scope='iteration')
+    def percent_race3(nodes_walk):
+        return np.log1p(nodes_walk.race_3_hhs / nodes_walk.households).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_race4', cache=True, cache_scope='iteration')
+    def percent_race4(nodes_walk):
+        return np.log1p(nodes_walk.race_4_hhs / nodes_walk.households).fillna(0)
+
+    @orca.column('nodes_walk', 'percent_hh_with_children', cache=True, cache_scope='iteration')
+    def percent_hh_with_children(nodes_walk):
+        return np.log1p(nodes_walk.hhs_with_children / nodes_walk.households).fillna(0)
+        
+    geographic_levels = [('parcels', 'parcel_id'),
+                         ('zones', 'zone_id'),
+                         ('nodes_walk', 'nodeid_walk'),
+                         ('nodes_drv', 'nodeid_drv')]
+
+    def make_disagg_var(from_geog_name, to_geog_name, var_to_disaggregate, from_geog_id_name, name_based_on_geography=True):
+        """
+        Generator function for disaggregating variables. Registers with orca.
+        """
+        if name_based_on_geography:
+            var_name = from_geog_name + '_' + var_to_disaggregate
+        else:
+            var_name = var_to_disaggregate
+        @orca.column(to_geog_name, var_name, cache=True, cache_scope='iteration')
+        def func():
+            print 'Disaggregating %s to %s from %s' % (var_to_disaggregate, to_geog_name, from_geog_name)
+
+            from_geog = orca.get_table(from_geog_name)
+            to_geog = orca.get_table(to_geog_name)
+            return misc.reindex(from_geog[var_to_disaggregate], to_geog[from_geog_id_name]).fillna(0)
+
+        return func
+                
+    for geography in geographic_levels:
+        geography_name = geography[0]
+        geography_id = geography[1]
+        if geography_name != 'buildings':
+            building_vars = orca.get_table('buildings').columns
+            for var in orca.get_table(geography_name).columns:
+                if var not in building_vars:
+                    make_disagg_var(geography_name, 'buildings', var, geography_id)
